Make JointBase responsible for its own tranmitting

+++ include/al5d_cpp/base/types.hpp
@@ -10,9 +10,9 @@ namespace al5d
 {
     typedef std::string Command;
     
-    typedef uint16_t PulseWidth;
-    typedef uint8_t BoardChannel;
-    typedef int16_t Degrees;
+    typedef unsigned int PulseWidth;
+    typedef unsigned int BoardChannel;
+    typedef unsigned int Degrees;
 }
